+++ src/pages/administrador/ModalesUsuarios/ModalCrear.jsx
@@ -1,53 +1,41 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import Modal from "react-bootstrap/Modal";
 import Form from "react-bootstrap/Form";
-import "../../pages/administrador/Administrador.css";
+import "../Administrador.css";
 import "bootstrap-icons/font/bootstrap-icons.css";
 import { useForm } from "react-hook-form";
-import useUsuarios from "../../stores/Usuarios-Store";
-import useSweetAlert from "../../hooks/useSweetAlert";
+import useUsuarios from "../../../stores/Usuarios-Store";
+import useSweetAlert from "../../../hooks/useSweetAlert";
 
-const ModalEditar = ({ usuario }) => {
+const ModalCrear = () => {
   const [show, setShow] = useState(false);
-  const { updateUsuario } = useUsuarios();
+  const { createUsuario } = useUsuarios();
   const { showAlert } = useSweetAlert();
   const {
     register,
     handleSubmit,
-    setValue,
     formState: { errors },
+    reset,
     watch,
   } = useForm();
-  const handleShow = () => setShow(true);
+
   const handleClose = () => {
     setShow(false);
+    reset();
   };
 
-  useEffect(() => {
-    if (usuario) {
-      setValue("rol", usuario.rol);
-      setValue("apellido", usuario.apellido);
-      setValue("nombre", usuario.nombre);
-      setValue("dni", usuario.dni);
-      setValue("telefono", usuario.telefono);
-      setValue("direccion", usuario.direccion);
-      setValue("email", usuario.email);
-      setValue("password", usuario.password);
-      setValue("passwordConfirm", usuario.passwordConfirm);
-      setValue("url_img", usuario.url_img);
-    }
-  }, [usuario, setValue]);
+  const handleShow = () => setShow(true);
 
   const onSubmit = async (data) => {
     try {
-      const respuesta = await updateUsuario(usuario.id, data);
+      const respuesta = await createUsuario(data);
       if (respuesta) {
-        await showAlert("success", "¡Listo!", "El usuario ha sido editado con éxito.");
+        await showAlert("success", "¡Listo!", "El usuario ha sido creado con éxito.");
       } else {
-        await showAlert("error", "Ocurrió un error", "No se pudo editar el usuario. Intenta nuevamente.");
+        await showAlert("error", "Ocurrió un error", "No se pudo crear el usuario. Intenta nuevamente.");
       }
     } catch (error) {
-      await showAlert("error", "Ocurrió un error", "No se pudo editar el usuario. Intenta nuevamente.");
+      await showAlert("error", "Ocurrió un error", "No se pudo crear el usuario. Intenta nuevamente.");
     }
     handleClose();
   };
@@ -56,14 +44,15 @@ const ModalEditar = ({ usuario }) => {
 
   return (
     <>
-      <button className="btn" onClick={handleShow}>
-        <i className="bi bi-pencil-square iconoEditar"></i>
+      <button className="p-2 rounded iconoCrear" onClick={handleShow}>
+        <i className="bi bi-file-earmark-plus-fill p-1"></i>
+        Nuevo usuario
       </button>
 
       <Modal show={show} onHide={handleClose} className="modalUsuario">
         <Modal.Header closeButton>
           <Modal.Title>
-            <span className="titulo">EDITAR USUARIO</span>{" "}
+            <span className="titulo">CREAR USUARIO</span>{" "}
           </Modal.Title>
         </Modal.Header>
         <Modal.Body>
@@ -263,7 +252,7 @@ const ModalEditar = ({ usuario }) => {
                     </Form.Text>
                   </Form.Group>
                   <Form.Group className="mb-3" controlId="formImagen">
-                    <Form.Label>URL de la Imagen</Form.Label>
+                    <Form.Label>URL de la imagen</Form.Label>
                     <Form.Control
                       type="text"
                       placeholder="Ingresar la URL de la imagen"
@@ -302,4 +291,4 @@ const ModalEditar = ({ usuario }) => {
   );
 };
 
-export default ModalEditar;
+export default ModalCrear;
